WEL (Win): Move the generated files in tools/wel and fix includes
