+++ tools/wel/osquery.man
