+++ tools/wel/osquery.rc
@@ -1,0 +1,3 @@
+LANGUAGE 0x9,0x1
+1 11 "tools/wel/MSG00001.bin"
+1 WEVT_TEMPLATE "tools/wel/osqueryTEMP.BIN"
