+++ tools/wel/osqueryTEMP.BIN
